support ini

+++ maintenance.py
@@ -11,6 +11,36 @@ import os
 import glob
 import getopt
 import sys
+import configparser
+import re
+
+config_ini_path = '/workdir/mount/config.ini'
+if os.path.exists(config_ini_path):
+    config_ini = configparser.ConfigParser()
+    config_ini.read(config_ini_path, encoding='utf-8')
+    config_default = config_ini['default']
+
+    if config_default.get('include_contract') is not None and config_default.get('include_contract') != '':
+        in_contract = re.compile(config_default.get('include_contract'))
+    else:
+        in_contract = None
+    if config_default.get('exclude_contract') is not None and config_default.get('exclude_contract') != '':
+        ex_contract = re.compile(config_default.get('exclude_contract'))
+    else:
+        ex_contract = None
+    if config_default.get('include_property') is not None and config_default.get('include_property') != '':
+        in_property = re.compile(config_default.get('include_property'))
+    else:
+        in_property = None
+    if config_default.get('exclude_property') is not None and config_default.get('exclude_property') != '':
+        ex_property = re.compile(config_default.get('exclude_property'))
+    else:
+        ex_property = None
+else:
+    in_contract = None
+    ex_contract = None
+    in_property = None
+    ex_property = None
 
 opts,args = getopt.getopt(sys.argv[1:],"xf")
 exec=False
@@ -39,18 +69,42 @@ if force:
     print('### DELETE Terraform STATE (rm -rf [path])')
     lgret = subprocess.check_output( [pwd + '/_scr/existTfstate.sh'] ).decode('ascii').splitlines()
     for v in lgret:
+        spl = v.split('/')
+        if in_contract is not None and in_contract.match(spl[4]) is None:
+            #print('A:SKIP: %s' % v)
+            continue
+        if ex_contract is not None and ex_contract.match(spl[4]) is not None:
+            #print('B:SKIP: %s' % v)
+            continue
+        if in_property is not None and in_property.match(spl[5]) is None:
+            #print('C:SKIP: %s' % v)
+            continue
+        if ex_property is not None and ex_property.match(spl[5]) is not None:
+            #print('D:SKIP: %s' % v)
+            continue
         if(exec):
             os.system('rm -rf '+v)
             print('DELETE: %s' % v)
         else:
             print(v)
         
-
 lgret = subprocess.check_output( [pwd + '/_scr/existPropVer.sh'] ).decode('ascii').splitlines()
 current={}
 for v in lgret:
     v=v.strip().split(':',2)
     spl=v[0].replace(pwd+'/','').split('/',3)
+    if in_contract is not None and in_contract.match(spl[1]) is None:
+        #print('A:SKIP: %s' % v)
+        continue
+    if ex_contract is not None and ex_contract.match(spl[1]) is not None:
+        #print('B:SKIP: %s' % v)
+        continue
+    if in_property is not None and in_property.match(spl[2]) is None:
+        #print('C:SKIP: %s' % v)
+        continue
+    if ex_property is not None and ex_property.match(spl[2]) is not None:
+        #print('D:SKIP: %s' % v)
+        continue
     current[spl[1]+'/'+spl[2]] = int(v[2].strip(', \''))
     
 lgret = json.loads(subprocess.check_output( ['akamai', 'pm', 'lg', '-f' ,'json','-s', 'default'] ))
@@ -62,9 +116,21 @@ remote={}
 # プロパティ関連のdiff
 for v in lgret:
     cid = v['contractIds'][0]
+    if in_contract is not None and in_contract.match(cid) is None:
+        #print('A:SKIP: %s' % v)
+        continue
+    if ex_contract is not None and ex_contract.match(cid) is not None:
+        #print('B:SKIP: %s' % v)
+        continue
     gid = v['groupId']
     lgret = json.loads(subprocess.check_output( ['akamai', 'pm', 'lpr', '-c', cid, '-g', gid, '-f' ,'json','-s', 'default'] ))
     for vv in lgret:
+        if in_property is not None and in_property.match(vv['propertyName']) is None:
+            #print('C:SKIP: %s' % vv)
+            continue
+        if ex_property is not None and ex_property.match(vv['propertyName']) is not None:
+            #print('D:SKIP: %s' % vv)
+            continue
         ver = 0
         if vv['productionVersion'] is not None and ver < vv['productionVersion']:
             ver = vv['productionVersion']
